ai/API/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.101.1
filelock==3.12.2
Flask==2.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
sklearn==0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7
zipp==3.16.2
